Cria crud

+++ deletar_salao.php
@@ -1,11 +1,13 @@
 <?php
     include "./SaloesDeFesta.php";
 
-    if(isset($_GET["id"])){
+    if(isset($_GET["localizacao"]) && isset($_GET["id"])){
         $id = $_GET["id"];
+        $localizacaoID = $_GET["localizacao"];
         $saloesDeFesta = new SaloesDeFesta();
-        $saloesDeFesta->delete($id);    
+        $saloesDeFesta->delete($id);
+
+        header("location: ./crud_saloes_disponiveis.php?localizacao=$localizacaoID");
     }
     
-    header("location: ./crud_saloes_disponiveis.php")
 ?>
